requirements.txt
fastapi[standard]
asyncpg
sqlalchemy
databases
psycopg2-binary
passlib
bcrypt==4.0.1
python-multipart
python-jose[cryptography]
